@@ -103,8 +103,9 @@
     "webpack-dev-server": "3.11.0"
   },
   "engines": {
-    "node": "12"
+    "node": "14"
   },
+  "type": "module",
   "private": true,
   "repository": {
     "type": "git",
