@@ -1,6 +1,4 @@
-'use strict';
-
-module.exports = [
+export const svgoOptions = [
   { addAttributesToSVGElement: false },
   { addClassesToSVGElement: false },
   { cleanupAttrs: false },
